requirements.txt
torch
torchvision
timm
open3d
numpy
opencv-python
matplotlib
pytorch-lightning
pyntcloud
scikit-learn
tqdm
requests
Pillow
yaml
einops
chamfer-distance
fastapi
uvicorn
python-multipart
pytorch3d
nvdiffrast